Refactor home tabs into separate components
